+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/tweaks/TweaksEvent.kt
@@ -45,4 +45,6 @@ sealed interface TweaksEvent : Event {
         override var canceled: Boolean = false,
     ) : TweaksEvent,
         CancelableEvent
+
+    object RenderHUDPre : TweaksEvent
 }