+++ src/main/kotlin/yqloss/yqlossclientmixinkt/module/tweaks/TweaksOptions.kt
@@ -24,4 +24,5 @@ interface TweaksOptions : YCModuleOptions {
     val enableInstantAim: Boolean
     val disablePearlClickBlock: Boolean
     val disableSkyBlockToolsNBTUpdateResetDigging: Boolean
+    val enableCatacombsBossBarFix: Boolean
 }